+++ src/auth/mod.rs
@@ -1,0 +1,4 @@
+pub mod login;
+pub mod signup;
+pub mod user_center;
+
