+++ src/lib.rs
@@ -1,0 +1,4 @@
+pub mod auth;
+pub mod db_lib;
+
+#[macro_use] extern crate rocket;
